test(systems): share stack constants instead of duplicating them
Cobalt and Beryl helpers now import chain IDs and wait_for_block from common. Each is path-included only by the binaries that call it, so unused helpers stay uncompiled.
