+++ etc/systems/tests/activation_registry.rs
@@ -1,5 +1,7 @@
 //! System tests for the activation registry precompile over Base node RPC.
 
+#[path = "common/beryl.rs"]
+mod beryl;
 #[path = "common/cobalt.rs"]
 mod cobalt;
 mod common;
@@ -15,13 +17,13 @@ use eyre::{Result, WrapErr};
 /// `isActivated` returns `false` for every feature id by default.
 #[tokio::test]
 async fn test_activation_registry_is_activated_default() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
 
     let output = client
         .call(
@@ -40,13 +42,13 @@ async fn test_activation_registry_is_activated_default() -> Result<()> {
 /// `admin()` returns the generated system test activation admin address.
 #[tokio::test]
 async fn test_activation_registry_admin() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let caller = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, caller.address()).await?;
+    beryl::wait_for_balance(&provider, caller.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &caller, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
 
     let output =
         client.call(ActivationRegistryStorage::ADDRESS, IActivationRegistry::adminCall {}).await?;
@@ -61,13 +63,13 @@ async fn test_activation_registry_admin() -> Result<()> {
 /// `setAdmin` stays disabled while Beryl is active but Cobalt is not.
 #[tokio::test]
 async fn test_activation_registry_set_admin_reverts_before_cobalt() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
 
     let succeeded = client
         .try_send_call(
@@ -91,13 +93,13 @@ async fn test_activation_registry_cobalt_admin_rotation() -> Result<()> {
         .wrap_err("Failed to parse system test admin private key")?;
     let new_admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_6.private_key)
         .wrap_err("Failed to parse system test new admin private key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
-    common::wait_for_balance(&provider, new_admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, new_admin.address()).await?;
 
     let admin_client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     let new_admin_client = B20PrecompileClient::new(&provider, &new_admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
 
     assert_eq!(admin_address(&admin_client).await?, admin.address());
 
@@ -142,13 +144,13 @@ async fn test_activation_registry_cobalt_admin_rotation() -> Result<()> {
 /// Admin activate/deactivate transitions update state, emit events, and reject repeated calls.
 #[tokio::test]
 async fn test_activation_registry_admin_lifecycle() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     let feature = ActivationFeature::B20Stablecoin.id();
 
     assert!(!is_activated(&client, feature).await?, "feature should start inactive");
@@ -209,13 +211,13 @@ async fn test_activation_registry_admin_lifecycle() -> Result<()> {
 /// Calling `activate` from a non-admin account reverts with `Unauthorized`.
 #[tokio::test]
 async fn test_activation_registry_unauthorized_activate_reverts() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let non_admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_6.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, non_admin.address()).await?;
+    beryl::wait_for_balance(&provider, non_admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &non_admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
 
     let succeeded = client
         .try_send_call(
@@ -244,13 +246,13 @@ async fn test_activation_registry_unauthorized_activate_reverts() -> Result<()> 
 /// `checkActivated` reverts before activation and succeeds while the feature is active.
 #[tokio::test]
 async fn test_activation_registry_check_activated_gate() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     let feature = ActivationFeature::B20Asset.id();
 
     assert!(
