+++ etc/systems/tests/common/cobalt.rs
@@ -1,20 +1,16 @@
-//! Cobalt stack helper. Not a `common` submodule so Beryl-only test binaries do not compile it.
+//! Cobalt stack helper. Path-included only by tests that activate Cobalt.
 
-use std::time::Duration;
-
-use alloy_provider::{Provider, RootProvider};
+use alloy_provider::RootProvider;
 use base_common_network::Base;
 use base_system_tests::{SystemTestStack, SystemTestStackBuilder};
-use eyre::{Result, WrapErr};
-use tokio::time::{sleep, timeout};
+use eyre::Result;
 
-const L1_CHAIN_ID: u64 = 1337;
-const L2_CHAIN_ID: u64 = 84538453;
-const BASE_AZUL_ACTIVATION_BLOCK: u64 = 0;
-const BASE_BERYL_ACTIVATION_BLOCK: u64 = 3;
-const BASE_COBALT_ACTIVATION_BLOCK: u64 = 5;
-const BLOCK_PRODUCTION_TIMEOUT: Duration = Duration::from_secs(30);
-const BLOCK_POLL_INTERVAL: Duration = Duration::from_millis(500);
+use super::common::{
+    BASE_AZUL_ACTIVATION_BLOCK, BASE_BERYL_ACTIVATION_BLOCK, L1_CHAIN_ID, L2_CHAIN_ID,
+    wait_for_block,
+};
+
+pub(crate) const BASE_COBALT_ACTIVATION_BLOCK: u64 = 5;
 
 /// Starts a system test stack with Cobalt active at block 5 and waits for block 6.
 ///
@@ -30,16 +26,6 @@ pub(crate) async fn start_cobalt_system() -> Result<(SystemTestStack, RootProvid
         .build()
         .await?;
     let provider = system.l2_builder_provider()?;
-    timeout(BLOCK_PRODUCTION_TIMEOUT, async {
-        loop {
-            let block = provider.get_block_number().await?;
-            if block > BASE_COBALT_ACTIVATION_BLOCK {
-                return Ok::<_, eyre::Error>(block);
-            }
-            sleep(BLOCK_POLL_INTERVAL).await;
-        }
-    })
-    .await
-    .wrap_err("Block production timed out")??;
+    wait_for_block(&provider, BASE_COBALT_ACTIVATION_BLOCK + 1).await?;
     Ok((system, provider))
 }
