+++ etc/systems/tests/in_process_zk.rs
@@ -2,6 +2,7 @@
 
 #[path = "common/cobalt.rs"]
 mod cobalt;
+mod common;
 
 use std::time::Duration;
