+++ etc/systems/tests/policy_registry.rs
@@ -1,5 +1,7 @@
 //! System tests for the policy registry precompile over Base node RPC.
 
+#[path = "common/beryl.rs"]
+mod beryl;
 mod common;
 
 use alloy_primitives::{Address, LogData};
@@ -13,13 +15,13 @@ use eyre::{Result, WrapErr};
 /// `createPolicy` emits the expected policy-registry events over RPC receipts.
 #[tokio::test]
 async fn test_policy_registry_create_policy_emits_events() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let caller = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, caller.address()).await?;
+    beryl::wait_for_balance(&provider, caller.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &caller, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     client.activate_feature(ActivationFeature::PolicyRegistry.id()).await?;
 
     let call = IPolicyRegistry::createPolicyCall {
@@ -60,13 +62,13 @@ async fn test_policy_registry_create_policy_emits_events() -> Result<()> {
 /// `policyExists(ALWAYS_ALLOW_ID)` returns `true` once the policy registry is active.
 #[tokio::test]
 async fn test_policy_registry_policy_exists() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let caller = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse system test private key")?;
-    common::wait_for_balance(&provider, caller.address()).await?;
+    beryl::wait_for_balance(&provider, caller.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &caller, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     client.activate_feature(ActivationFeature::PolicyRegistry.id()).await?;
 
     let output = client
@@ -86,18 +88,18 @@ async fn test_policy_registry_policy_exists() -> Result<()> {
 /// Custom policy creation, membership, admin transfer, renounce, and error paths round-trip over RPC.
 #[tokio::test]
 async fn test_policy_registry_lifecycle_and_error_paths() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse policy admin key")?;
     let next_admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_6.private_key)
         .wrap_err("Failed to parse next policy admin key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
-    common::wait_for_balance(&provider, next_admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, next_admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     let next_admin_client = B20PrecompileClient::new(&provider, &next_admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     client.activate_feature(ActivationFeature::PolicyRegistry.id()).await?;
 
     let allowlist_id = create_policy(
@@ -253,13 +255,13 @@ async fn test_policy_registry_lifecycle_and_error_paths() -> Result<()> {
 /// View calls remain available while write calls are gated when the policy registry is deactivated.
 #[tokio::test]
 async fn test_policy_registry_deactivated_views_and_write_gate() -> Result<()> {
-    let (_system, provider) = common::start_beryl_system().await?;
+    let (_system, provider) = beryl::start_beryl_system().await?;
     let admin = PrivateKeySigner::from_bytes(&ANVIL_ACCOUNT_5.private_key)
         .wrap_err("Failed to parse policy admin key")?;
-    common::wait_for_balance(&provider, admin.address()).await?;
+    beryl::wait_for_balance(&provider, admin.address()).await?;
 
     let client = B20PrecompileClient::new(&provider, &admin, common::L2_CHAIN_ID)
-        .with_receipt_timeout(common::TX_RECEIPT_TIMEOUT);
+        .with_receipt_timeout(beryl::TX_RECEIPT_TIMEOUT);
     client.activate_feature(ActivationFeature::PolicyRegistry.id()).await?;
 
     let blocklist_id = create_policy(
